--- a/queue/queue.go
+++ b/queue/queue.go
@@ -160,8 +160,13 @@ func (q *Queue) Put(items ...interface{}) error {
 			break
 		}
 		sema.response.Add(1)
-		sema.ready <- true
-		sema.response.Wait()
+		select {
+		case sema.ready <- true:
+			sema.response.Wait()
+		default:
+			// This semaphore timed out.
+			sema.response.Done()
+		}
 		if len(q.items) == 0 {
 			break
 		}