This does a few things:
- Add a `Poll` method to `Queue` which allows you to perform a `Get` with a timeout.
- Export `ErrDisposed` to allow users to check error type.
- Export `ErrTimeout` to allow users to check for timeout vs disposal.
- The internal `sema` struct now uses a channel instead of a `sync.WaitGroup` to provide support for timeouts. The effect on performance is negligible (see below).

Before:
```
BenchmarkPriorityQueue   3000000               517 ns/op
BenchmarkQueue  10000000               194 ns/op
BenchmarkChannel         3000000               672 ns/op
BenchmarkQueuePut          10000            141664 ns/op
BenchmarkQueueGet          10000            112553 ns/op
BenchmarkExecuteInParallel         20000             82021 ns/op
BenchmarkRBLifeCycle      500000              3142 ns/op
BenchmarkRBPut  20000000                85.2 ns/op
BenchmarkRBGet  50000000               484 ns/op
```

After:
```
PASS
BenchmarkPriorityQueue   3000000               539 ns/op
BenchmarkQueue  10000000               197 ns/op
BenchmarkChannel         3000000               574 ns/op
BenchmarkQueuePut          10000            125814 ns/op
BenchmarkQueueGet          10000            104068 ns/op
BenchmarkQueuePoll         20000            109937 ns/op
BenchmarkExecuteInParallel         30000             62881 ns/op
BenchmarkRBLifeCycle      500000              3198 ns/op
BenchmarkRBPut  20000000               101 ns/op
BenchmarkRBGet  50000000               473 ns/op
```

@dustinhiatt-wf @alexandercampbell-wf @stevenosborne-wf @beaulyddon-wf @tannermiller-wf