@tylertreat-wf 2015-06-25
There appears to be a race condition somewhere causing a deadlock:

```
go test -bench=BenchmarkQueue -run=XXX ./queue
PASS
BenchmarkQueue-8        fatal error: all goroutines are asleep - deadlock!

goroutine 1 [chan receive]:
testing.(*B).run(0xc208094600, 0x0, 0x0, 0x0, 0x0, 0x0)
        /usr/local/go/src/testing/benchmark.go:180 +0x67
testing.RunBenchmarks(0x3b03f8, 0x49e5a0, 0xa, 0xa)
        /usr/local/go/src/testing/benchmark.go:312 +0x649
testing.(*M).Run(0xc2080683c0, 0x4a9140)
        /usr/local/go/src/testing/testing.go:494 +0x1a5
main.main()
        github.com/Workiva/go-datastructures/queue/_test/_testmain.go:138 +0x1d5

goroutine 5 [semacquire]:
sync.(*WaitGroup).Wait(0xc208022000)
        /usr/local/go/src/sync/waitgroup.go:132 +0x169
github.com/Workiva/go-datastructures/queue.BenchmarkQueue(0xc208094600)
        /Users/tylertreat/Go/src/github.com/Workiva/go-datastructures/queue/queue_test.go:267 +0x37d
testing.(*B).runN(0xc208094600, 0x2710)
        /usr/local/go/src/testing/benchmark.go:124 +0x95
testing.(*B).launch(0xc208094600)
        /usr/local/go/src/testing/benchmark.go:216 +0x159
created by testing.(*B).run
        /usr/local/go/src/testing/benchmark.go:179 +0x3e

goroutine 8 [select]:
github.com/Workiva/go-datastructures/queue.(*Queue).Poll(0xc2080aa040, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0, 0x0)
        /Users/tylertreat/Go/src/github.com/Workiva/go-datastructures/queue/queue.go:215 +0x5b9
github.com/Workiva/go-datastructures/queue.(*Queue).Get(0xc2080aa040, 0x1, 0x0, 0x0, 0x0, 0x0, 0x0)
        /Users/tylertreat/Go/src/github.com/Workiva/go-datastructures/queue/queue.go:183 +0x68
github.com/Workiva/go-datastructures/queue.func·012()
        /Users/tylertreat/Go/src/github.com/Workiva/go-datastructures/queue/queue_test.go:254 +0x58
created by github.com/Workiva/go-datastructures/queue.BenchmarkQueue
        /Users/tylertreat/Go/src/github.com/Workiva/go-datastructures/queue/queue_test.go:261 +0x251
exit status 2
FAIL    github.com/Workiva/go-datastructures/queue      0.012s
```