@tylertreat-wf 2015-06-25
Fixed the deadlock and addressed comments.

@dustinhiatt-wf @alexandercampbell-wf