@alexandercampbell-wf 2015-06-26
+1

Unrelated, but maybe we should add `go test -bench . -benchtime 1ms` to CI and integration tests, just to confirm the benchmarks have no panics.